Fix session key security issues
Since the password was set to None, the session hash was always
identical and predictable for an attacker. A new random salt is
added to replace the password which served this funciton before.

Should the new session salt is set be default to a rendom value.
Should the salt be set to None for some reason, the
`get_session_auth_hash` method will raise a `ValueError`.

The password field is now removed from the user model. It will
raise a `FieldDoesNotExist` error, should the attribute be
access further preventing similar security issues.