Switch token separate to URL safe character . (dot)
Some email clients replace double slashes with a single slash.
The double slash occured for users with no last_login date (newly
created users).

To bypass this issue, the separator is changed to . (dot) as it is
a non-reserved URL safe caracter (RFC3986 2.3) and not port of the
base64url alphabet.

See also:
https://www.ietf.org/rfc/rfc3986.txt
https://tools.ietf.org/html/rfc4648