Drop custom separator in favor of default colon (`:`)
Dot is unreserved character, however the representation in browsers (`%2E` or `.`), mail clients and even SMTP-servers is not unified and can lead to different issues:
- https://wordtothewise.com/2018/11/why-do-my-urls-have-two-dots/
- https://bugs.python.org/issue43922

Also, Django is using dot for some funky JSON compression in `django.core.signing`:
https://github.com/django/django/blob/4f8c7fd9d91b35e2c2922de4bb50c8c8066cbbc6/django/core/signing.py#L29-L30

So, in this PR I am suggesting to drop custom separator and switch to the default colon (`:`).
It is also not reserved of anything in domain path, but only in host and scheme parts.
- https://www.rfc-editor.org/rfc/rfc3986#section-3.3
- https://security.stackexchange.com/questions/159099/colons-in-urls-safe